update: I don't know, but lots of updates

+++ types/mempool.go
@@ -3,5 +3,5 @@ package types
 type IMempool interface {
 	Lock()
 	Unlock()
-	Update(height int, txs []Tx)
+	Update(height int64, txs []Tx)
 }
